requirements.txt
streamlit
opencv-python-headless
ultralytics
twilio
numpy
pandas
python-dotenv
